reorganisation of server
