+++ Makefile
@@ -26,10 +26,8 @@ test-build:
 test-server: export FT_NEXT_BACKEND_KEY=test-backend-key
 test-server: export FT_NEXT_BACKEND_KEY_OLD=test-backend-key-old
 test-server: export FT_NEXT_BACKEND_KEY_OLDEST=test-backend-key-oldest
-test-server: export LAYOUT_POLLING_INTERVAL=100
 test-server:
 	mocha node/test/app.test.js node/test/**/*.test.js  --recursive
-
 
 nightwatch:
 	nht nightwatch test/js-success.nightwatch.js
