+++ node/lib/hashed-assets.js
