+++ node/lib/n-ui-manager.js
