--- node/src/anon/index.js
@@ -1,3 +1,0 @@
-module.exports = {
-	middleware : require('./middleware')
-};
